MCA/base: Add new MCA variable system
Features:
 - Support for an override parameter file (openmpi-mca-param-override.conf).
   Variable values in this file can not be overridden by any file or environment
   value.
 - Support for boolean, unsigned, and unsigned long long variables.
 - Support for true/false values.
 - Support for enumerations on integer variables.
 - Support for MPIT scope, verbosity, and binding.
 - Support for command line source.
 - Support for setting variable source via the environment using
   OMPI_MCA_SOURCE_<var name>=source (either command or file:filename)
 - Cleaner API.
 - Support for variable groups (equivalent to MPIT categories).

Notes:
 - Variables must be created with a backing store (char **, int *, or bool *)
   that must live at least as long as the variable.
 - Creating a variable with the MCA_BASE_VAR_FLAG_SETTABLE enables the use of
   mca_base_var_set_value() to change the value.
 - String values are duplicated when the variable is registered. It is up to
   the caller to free the original value if necessary. The new value will be
   freed by the mca_base_var system and must not be freed by the user.
 - Variables with constant scope may not be settable.
 - Variable groups (and all associated variables) are deregistered when the
   component is closed or the component repository item is freed. This
   prevents a segmentation fault from accessing a variable after its component
   is unloaded.
 - After some discussion we decided we should remove the automatic registration
   of component priority variables. Few component actually made use of this
   feature.
 - The enumerator interface was updated to be general enough to handle
   future uses of the interface.
 - The code to generate ompi_info output has been moved into the MCA variable
   system. See mca_base_var_dump().

opal: update core and components to mca_base_var system
orte: update core and components to mca_base_var system
ompi: update core and components to mca_base_var system

This commit also modifies the rmaps framework. The following variables were
moved from ppr and lama: rmaps_base_pernode, rmaps_base_n_pernode,
rmaps_base_n_persocket. Both lama and ppr create synonyms for these variables.

This commit was SVN r28236.

+++ ompi/mca/bcol/iboffload/bcol_iboffload_module.c
@@ -768,7 +768,7 @@ static inline struct ibv_cq *ibv_create_cq_compat(struct ibv_context *context,
 int mca_bcol_iboffload_adjust_cq(mca_bcol_iboffload_device_t *device,
                                                    struct ibv_cq **ib_cq)
 {
-    uint32_t cq_size = mca_bcol_iboffload_component.cq_size;
+    uint32_t cq_size = (uint32_t) mca_bcol_iboffload_component.cq_size;
 
     if (NULL == *ib_cq) {
         *ib_cq = ibv_create_cq_compat(device->dev.ib_dev_context, cq_size,
@@ -1070,8 +1070,8 @@ mca_bcol_iboffload_comm_query(mca_sbgp_base_module_t *sbgp, int *num_modules)
         }
 
         memset(&mqe_attr, 0, sizeof(mqe_attr));
-        mqe_attr.max_mqe_tasks = mca_bcol_iboffload_component.max_mqe_tasks;
-        mqe_attr.max_mq_size = mca_bcol_iboffload_component.max_mq_size;
+        mqe_attr.max_mqe_tasks = (uint32_t)mca_bcol_iboffload_component.max_mqe_tasks;
+        mqe_attr.max_mq_size = (uint32_t)mca_bcol_iboffload_component.max_mq_size;
         mqe_attr.cq = iboffload_module->device->ib_mq_cq;
 
         /* ALL MQs have the same configuration */
